+++ examples/test_simple_break.py
@@ -1,0 +1,5 @@
+def test_simple_break():
+    """Test basic break statement"""
+    for i in range(5):
+        if i == 3:
+            break